refactor(vignettes): moved analysis out of vigenettes - see explanation...
When structuring research as package, it is often recommended for the analysis to be performed as vignettes. However, this causes errors for the package when trying to save files. I don't see any benefit from putting analysis in vignettes/ when it causes issues like this, so have moved to a different folder with a more explanative name (notebooks/)